\name{CoCoGraph-internal}
\title{Internal CoCoGraph objects}
\alias{.First.lib}
\alias{.First.lib.CoCoDynamicGraph}
\alias{.onAttach}
\alias{.onLoad}
\alias{.packageName}

\alias{.--C--CoCoTestClass}
\alias{.--M--dynamic.Graph:.GlobalEnv}
\alias{.--M--label:.GlobalEnv}
\alias{.--M--modifyModel:.GlobalEnv}
\alias{.--M--subModifyModel:.GlobalEnv}
\alias{.--M--testEdge:.GlobalEnv}
\alias{.--M--width:.GlobalEnv}

\alias{label} % CoCoTestClass-method
\alias{width} % CoCoTestClass-method

\alias{dynamic.Graph,CoCoModelClass-method}
\alias{dynamic.Graph,numeric-method}
\alias{dynamic.Graph,character-method}
\alias{modifyModel,CoCoModelClass-method}
\alias{subModifyModel,CoCoModelClass-method}
\alias{testEdge,CoCoModelClass-method}

\alias{dynamic.Graph} % CoCoModelClass-method
\alias{modifyModel} % CoCoModelClass-method
\alias{subModifyModel} % CoCoModelClass-method
\alias{testEdge} % CoCoModelClass-method

\description{Internal CoCoGraph objects.}
\details{These are not to be called by the user.}
\keyword{internal}
