\name{newCoCoTestObject}
\alias{newCoCoTestObject}

\title{ Return a CoCo test object }

\description{ Return a CoCo test object }

\usage{
newCoCoTestObject(test)
}

\arguments{
  \item{test}{ A numeric vecttor, the result of \code{\link[CoCoRaw]{returnDeviance}}. }
}

% \details{ }

% \value{ }

% \references{Badsberg, J.H.: Xlisp+CoCo, Aalborg University, 1996. }

\author{ Jens Henrik Badsberg }

% \note{ }

\seealso{ \code{\link[CoCo]{makeCoCo}}. }

\keyword{ methods }% at least one, from doc/KEYWORDS
\keyword{ multivariate }% __ONLY ONE__ keyword per line

