\name{CoCoGraph-package}
\alias{CoCoGraph-package}
\docType{package}
\title{CoCoGraph}
\description{
  The package provides an graphical user interface
  to graphs for models in CoCo and CoCoCg.
  The linked views of graphs are presented and manipulated by Tcl/tk.
}
\details{
   \tabular{ll}{
      Package: \tab CoCoGraph    \cr
      Type:    \tab Package      \cr
      Version: \tab 0.1.7.6      \cr
      Date:    \tab 2010-03-17   \cr
   }
}
\author{
Jens Henrik Badsberg

Maintainer: Jens Henrik Badsberg <coco@badsberg.eu>
}
% \references{ }
\keyword{ package }
